feat(roles): Vendors solr and java roles
I pulled these roles from Github and vendored them here. Ansible has
tricks for using existing roles like these through their "galaxy" tool,
but:

1. It's too danged complicated, and
2. I don't trust stuff on other people's computers.

Re #2, the point is that I feel really weird about installing a ton of
stuff from the Ansible galaxy that has no vetting or anything of the
type. I guess that's what we do with pip, and to an extent with apt, but
somehow it feels different when provisioning a server. I'm probably
wrong about that.

Anyway, this method makes it easy to have a solr.yml file that plays a
bunch of roles, so it's convenient from that perspective too. Watch for
that file shortly.