feat(config): Add solr configurations
This is a big commit with a ton of duplication. What we have here is the
default solr 4.x solr configuration directory, replicated once per core.

Yeah, it's not ideal, but I actually kind of like it. It's *simple*.
Later, if we get more sophisticated or if somebody feels like it, they
can identify the differences between these directories (I think it's
only the schema files at this point), but for now this is *fine*.

Let's not let perfection be the enemy of progress, eh?